+++ src/facade.rs
@@ -1,0 +1,5 @@
+pub mod frontend_api {
+  pub fn ciao_amanda () {
+    println!("Ciao Amanda");
+  }
+}
